+++ crates/directory/src/core/dispatch.rs
@@ -5,6 +5,7 @@
  */
 
 use crate::{Account, Credentials, Directory, Recipient, backend::oidc::OidcDiscovery};
+use registry::schema::enums::DirectoryType;
 use trc::AddContext;
 
 impl Directory {
@@ -13,6 +14,7 @@ impl Directory {
             Directory::Ldap(store) => store.authenticate(credentials).await,
             Directory::Sql(store) => store.authenticate(credentials).await,
             Directory::OpenId(store) => store.authenticate(credentials).await,
+            Directory::Unavailable(directory) => Err(directory.error()),
         }
         .caused_by(trc::location!())
     }
@@ -22,16 +24,25 @@ impl Directory {
             Directory::Ldap(store) => store.recipient(address).await,
             Directory::Sql(store) => store.recipient(address).await,
             Directory::OpenId(_) => Ok(Recipient::Invalid), // OIDC directories do not support recipient lookups
+            Directory::Unavailable(directory) => Err(directory.error()),
         }
         .caused_by(trc::location!())
     }
 
     pub fn has_bearer_token_support(&self) -> bool {
-        matches!(self, Directory::OpenId(_))
+        match &self {
+            Directory::OpenId(_) => true,
+            Directory::Unavailable(directory) => directory.directory_type() == DirectoryType::Oidc,
+            _ => false,
+        }
     }
 
     pub fn can_lookup_recipients(&self) -> bool {
-        !matches!(self, Directory::OpenId(_))
+        match &self {
+            Directory::OpenId(_) => false,
+            Directory::Unavailable(directory) => directory.directory_type() != DirectoryType::Oidc,
+            _ => true,
+        }
     }
 
     pub fn oidc_discovery_document(&self) -> Option<&OidcDiscovery> {