+++ tests/src/directory/mod.rs
@@ -11,11 +11,13 @@ pub mod oidc;
 #[cfg(feature = "sqlite")]
 pub mod sql;
 pub mod synchronization;
+pub mod unavailable;
 
 #[tokio::test(flavor = "multi_thread")]
 pub async fn directory_tests() {
     ldap::test().await;
     oidc::test().await;
+    unavailable::test().await;
     discovery::test().await;
     #[cfg(feature = "sqlite")]
     sql::test().await;